Project6_Sidekick/SideKick/requirements.txt
langchain>=0.1.0
langchain_community>=0.0.20
langchain-groq>=0.0.5
langchain_core>=0.1.0
langgraph>=0.0.40
python-dotenv>=1.0.0
requests>=2.31.0
gradio>=4.0.0
typing_extensions>=4.8.0
pydantic>=2.0.0
IPython>=8.0.0